agro lake/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seagri-server"
version = "0.1.0"
description = "MCP Server para consulta de dados agrícolas e gerenciamento de operações com integração Apidog"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.20.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "customtkinter>=5.2.0",
    "google-generativeai>=0.8.3",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
]

[project.scripts]
seagri-server = "server:main"
seagri-client = "client:main"

[tool.setuptools]
packages = ["gui", "models", "services", "utils"]

[tool.setuptools.package-dir]
"" = "."
